refact(*): Race condition during notify fixed
When the mouse enters and leave the widget in a short amount of time,
the async function finishes the execution only after destroy was
invoked. Which means that an orphan notification stays on the screen and
has to be manually closed or reinvoked properly.

To avoid this race condition a default timeout of 2 seconds was added.

Besides, some redundant code got cleaned during the process.